Attack animation + damage + Dummy
Added the animation for the attack.
2 animation to add a bit of variety .
Damage variable created and a dummy has been created to test everything.  Dummy can be cut in half and has a certain amount of life. Also knockback is added when the player hit the dummy.

Next thing to do are changing the dummy into a proper enemy that can move and attack the player.