Enemy State Machine and Other Stuff
Creating the first enemy (not counting the dummy made for testing).

Started by creating the different animation to use Walking and Knockback when hit animation.
Then created a state machine for the enemy for the Walking, Knockback and Dead state.
Added hitbox collider etc..
Put everything on the correct layer.
Made a Prefab of the enemy
added particle for blood splatter when hit and when the enemy is killed.

made some fix where the player could not hit the enemy, now everything working fine.
Enemy is able to detect the edge and wall, so he act like a koopa in super Mario by walking back and forth.

Next his adding attack to the enemy that damage the player and maybe respawn.